Seperated the custom stylings and the data out to their own respective functional and presentational components
